Started framework for real code.
Its quite different from the pseudo code but we can change it later. I was trying something but I don't know if it will work. My reason behind the change is that Davids code requires us knowing certain values instead of the robot figuring out the values through the picture which it needs to do.